requirements.txt

aiogram==3.18.0
aiogram_dialog==2.3.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
aiosqlite==0.20.0
annotated-types==0.7.0
asyncpg    == 0.30.0
attrs         ==    25.1.0
cachetools      ==  5.5.2
certifi           ==2025.1.31
environs          ==14.1.1
frozenlist    ==    1.5.0
idna            ==  3.10
iso8601         ==  2.1.0
Jinja2          ==  3.1.5
magic-filter    ==  1.0.12
MarkupSafe      ==  3.0.2
marshmallow     ==  3.26.1
multidict       ==  6.1.0
packaging      ==    24.2
pip              ==  24.0
propcache      ==    0.3.0
pydantic       ==    2.10.6
pydantic_core  ==    2.27.2
pypika-tortoise ==   0.5.0
python-dotenv  ==    1.0.1
pytz          ==     2025.1
tortoise-orm   ==    0.24.1
typing_extensions==  4.12.2
yarl      ==         1.18.3